Release 1.9.1. For changelog, check CHANGELOG.rst

+++ ask-smapi-model/ask_smapi_model/v1/audit_logs/request_pagination_context.py
@@ -33,12 +33,12 @@ class RequestPaginationContext(object):
     :param next_token: When the response to this API call is truncated, the response includes the nextToken element. The value of nextToken can be used in the next request as the continuation-token to list the next set of objects. The continuation token is an opaque value that this API understands. Token has expiry of 1 hour.
     :type next_token: (optional) str
     :param max_results: Sets the maximum number of results returned in the response body. If you want to retrieve more or less than the default of 50 results, you can add this parameter to your request. maxResults can exceed the upper limit of 250 but we will not return more items than that. The response might contain fewer results than maxResults for purpose of keeping SLA or because there are not enough items, but it will never contain more.
-    :type max_results: (optional) float
+    :type max_results: (optional) int
 
     """
     deserialized_types = {
         'next_token': 'str',
-        'max_results': 'float'
+        'max_results': 'int'
     }  # type: Dict
 
     attribute_map = {
@@ -48,13 +48,13 @@ class RequestPaginationContext(object):
     supports_multiple_types = False
 
     def __init__(self, next_token=None, max_results=None):
-        # type: (Optional[str], Optional[float]) -> None
+        # type: (Optional[str], Optional[int]) -> None
         """This object includes nextToken and maxResults.
 
         :param next_token: When the response to this API call is truncated, the response includes the nextToken element. The value of nextToken can be used in the next request as the continuation-token to list the next set of objects. The continuation token is an opaque value that this API understands. Token has expiry of 1 hour.
         :type next_token: (optional) str
         :param max_results: Sets the maximum number of results returned in the response body. If you want to retrieve more or less than the default of 50 results, you can add this parameter to your request. maxResults can exceed the upper limit of 250 but we will not return more items than that. The response might contain fewer results than maxResults for purpose of keeping SLA or because there are not enough items, but it will never contain more.
-        :type max_results: (optional) float
+        :type max_results: (optional) int
         """
         self.__discriminator_value = None  # type: str
